# Centralize RYU_32_BIT_PLATFORM and optimize shiftright128().